refactor: move routers and controllers to its own files
