+++ src/test/unit/controllers/claimTaskController.test.ts
@@ -1,0 +1,55 @@
+import { Response } from 'express';
+import { claimTask } from '../../../main/controllers/claimTaskController';
+
+describe('claimTask controller', () => {
+  /* eslint-disable  @typescript-eslint/no-explicit-any */
+  let req = {} as any;
+  const res = {} as Response;
+
+  beforeEach(() => {
+
+    res.render = jest.fn();
+
+  });
+
+  test('claimTask method', () => {
+
+    req = (
+      {
+        query: {
+          caseRef: '3',
+        },
+        session: {
+          myTasks: [{ caseRef: '1' }, { caseRef: '2' }],
+          myAvailableTasks: [{ caseRef: '3' }, { caseRef: '4' }],
+          myFilteredAvailableTasks: [{ caseRef: '3' }, { caseRef: '4' }],
+          addLocations: [
+            {
+              index: 1,
+              name: 'Birmingham',
+            },
+          ],
+          removeLocations: [],
+        },
+        /* eslint-disable  @typescript-eslint/no-explicit-any */
+      } as any);
+
+    claimTask(req, res);
+
+    expect(res.render).toHaveBeenCalledTimes(1);
+    expect(res.render).toHaveBeenCalledWith('task-list', {
+      tasks: {
+        myAvailableTasks: [{ caseRef: '4' }],
+        myTasks: [{ caseRef: '1' }, { caseRef: '2' }, { caseRef: '3' }],
+        addLocations: [
+          {
+            index: 1,
+            name: 'Birmingham',
+          },
+        ],
+        removeLocations: [],
+      },
+    });
+  });
+
+});
